requirements.txt
# Automated Metadata Generation System Requirements
# ================================================
# Core dependencies for the document processing system

# Web framework
streamlit>=1.28.0

# Document processing
pymupdf>=1.23.0          # PDF processing and text extraction
python-docx>=0.8.11      # Microsoft Word document processing

# OCR (Optical Character Recognition)
pytesseract>=0.3.10      # Python wrapper for Tesseract OCR
Pillow>=10.0.0           # Image processing for OCR

# Text processing and NLP
# Note: We're using simple, built-in approaches to keep it beginner-friendly
# No heavy NLP libraries needed for this basic version

# Data handling
pandas>=1.5.0            # For potential data manipulation (optional)

# Development and testing (optional)
pytest>=7.0.0            # For running tests
jupyter>=1.0.0           # For the Jupyter notebook version

# Additional utilities
python-dateutil>=2.8.2   # Better date parsing (optional enhancement)
